+++ src/Show.hs
@@ -9,8 +9,6 @@ module Show
        ( Print (..)
        , putText
        , putLText
-       , putByteString
-       , putLByteString
        ) where
 
 import qualified Base
@@ -59,11 +57,3 @@ putText = putStrLn
 putLText :: MonadIO m => TL.Text -> m ()
 putLText = putStrLn
 {-# SPECIALIZE putLText :: TL.Text -> Base.IO () #-}
-
-putByteString :: MonadIO m => BS.ByteString -> m ()
-putByteString = putStrLn
-{-# SPECIALIZE putByteString :: BS.ByteString -> Base.IO () #-}
-
-putLByteString :: MonadIO m => BL.ByteString -> m ()
-putLByteString = putStrLn
-{-# SPECIALIZE putLByteString :: BL.ByteString -> Base.IO () #-}